% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRGB.R
\name{plotRGB}
\alias{plotRGB}
\title{plotRGB}
\usage{
plotRGB(rgb, max_val = 1)
}
\arguments{
\item{rgb}{a numeric vector of length 3 of rgbs scaled from 0-255}
}
\description{
method to quickly visualize an RGB color
}
