% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraitsOnTree.R
\name{plotTraitOnTree}
\alias{plotTraitOnTree}
\title{Plot a phytools contMap of a trait}
\usage{
plotTraitOnTree(
  df,
  trait_colname,
  tree_location,
  replace_underscores = T,
  remove_commas = T,
  remove_first_split = F
)
}
\arguments{
\item{df}{df containing trait}

\item{trait_colname}{colname for the trait}

\item{tree_location}{file loc of the tree (read using ape::read.tree)}

\item{replace_underscores}{whether underscores should be replaced with spaces in the loaded tree}
}
\value{
nothing
}
\description{
Plot a phytools contMap of a trait
}
